Added Fonts, Changed to PHP for header addin
