+++ index.php
@@ -7,43 +7,12 @@
 	<link rel="stylesheet" href="assets/css/main.css">
 	<script src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.4/jquery.min.js"></script>
 	<script src="assets/javascript/ajax_dynamic.js"></script>
-	<script>
-		var timer_out = setTimeout(function(){
-			if ($("#loading").is(":visible")) {
-				$("#loading").hide()
-			}
-		}, 15000)
-	</script>
 	<meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
 </head>
 <body>
-	<div id="loading">
-		<div id="shadow"></div>
-		<div id="diag">
-			<h1>Loading</h1>
-			<p>Please wait while we load the document</p>
-			<img src="http://i.imgur.com/6EkrPfM.gif?1" style="display:none;" id="load-spinner" title="Loading..." alt="Loading Animation" />
-		</div>
-	</div>
-	<script>
-		$("#load-spinner").load(function(){
-			$("#load-spinner").slideDown(250)
-		})
-		$("#loading").show()
-		$(window).load(function(){
-			setTimeout(function() { $("#loading").slideUp(500) } , 500)
-			clearTimeout(timer_out)
-		})
-	</script>
-	<header><!--This header is fixed, meaning it will stay at the top of the page constantly, a media query will disable this on a mobile level-->
-		<nav>
-			<ul>
-				<a href="index.html" class="ajax_load active"><li>Home</li></a>
-				<a href="#" class="ajax_load"><li>About</li></a>
-				<a href="#" class="ajax_load"><li>Contact</li></a>
-			</ul>
-		</nav>
-	</header>
+	<?php
+		require_once"assets/server/header_addin.php";
+	?>
 	<div class="page-bg" id="main-page-bg"> <!--Faded out and replaced using ajax--> </div>
 	<div class="page-container current" id="index"> <!-- Slide off screen, slide new page on screen using ajax -->
 		<div id="wrapper">
@@ -73,7 +42,7 @@
 								<div class="game-title"><span>Garrys Mod</span></div></a>
 						</div>
 						<div class="game">
-							<a href="csgo.html" class="ajax_load"><div class="game-bg" id="CS"></div>
+							<a href="csgo.php" class="ajax_load"><div class="game-bg" id="CS"></div>
 								<div class="game-title"><span>Counter Strike: Global Offensive</span></div></a>
 						</div>
 						<div class="game">
